+++ packages/webcodecs/src/get-config.ts
@@ -21,27 +21,3 @@ export const getVideoDecoderConfigWithHardwareAcceleration = async (
 
 	return null;
 };
-
-export const getVideoEncoderConfigWithHardwareAcceleration = async (
-	config: VideoEncoderConfig,
-): Promise<VideoEncoderConfig | null> => {
-	const hardware: VideoEncoderConfig = {
-		...config,
-		hardwareAcceleration: 'prefer-hardware',
-	};
-
-	if ((await VideoEncoder.isConfigSupported(hardware)).supported) {
-		return hardware;
-	}
-
-	const software: VideoEncoderConfig = {
-		...config,
-		hardwareAcceleration: 'prefer-software',
-	};
-
-	if ((await VideoEncoder.isConfigSupported(software)).supported) {
-		return software;
-	}
-
-	return null;
-};
